requirements-dev.txt
django-browser-reload==1.18.0
django-extensions==4.1
pytest==8.*
pytest-django==4.*
ruff==0.7.*
black==24.*
